feature(babel): added babel - reorganized folders - fixed paths - adjusted tests to inject compiled files from lib/ folder - adjusted build script
