+++ .babelrc
@@ -1,0 +1,15 @@
+{
+    "presets": [
+        [
+            "@babel/preset-env",
+            {
+                "targets": {
+                    "browsers": [
+                        "last 1 versions",
+                        "ie >= 11"
+                    ]
+                }
+            }
+        ]
+    ]
+}
