+++ assets/img/clippy-icon-48x48.png
