+++ manifest.json
@@ -7,42 +7,42 @@
     "version": "1.9.0",
     "author": "Ante Barić (capJavert)",
     "icons": {
-        "16": "src/assets/img/clippy-icon-16x16.png",
-        "48": "src/assets/img/clippy-icon-48x48.png",
-        "128": "src/assets/img/clippy-icon-128x128.png"
+        "16": "assets/img/clippy-icon-16x16.png",
+        "48": "assets/img/clippy-icon-48x48.png",
+        "128": "assets/img/clippy-icon-128x128.png"
     },
     "browser_action": {
         "default_title": "Clippy Assistant",
         "default_icon": {
-            "16": "src/assets/img/clippy-icon-gray-48x48.png",
-            "24": "src/assets/img/clippy-icon-gray-48x48.png",
-            "32": "src/assets/img/clippy-icon-gray-48x48.png"
+            "16": "assets/img/clippy-icon-gray-48x48.png",
+            "24": "assets/img/clippy-icon-gray-48x48.png",
+            "32": "assets/img/clippy-icon-gray-48x48.png"
         }
     },
     "content_scripts": [
         {
             "matches": ["<all_urls>"],
             "css": [
-                "src/assets/css/clippy.css"
+                "assets/css/clippy.css"
             ],
             "js": [
-                "src/assets/js/jquery.min.js",
-                "src/assets/js/clippy.js",
-                "src/assets/js/agent.js",
-                "src/index.js"
+                "assets/js/jquery.min.js",
+                "assets/js/clippy.js",
+                "assets/js/agent.js",
+                "lib/index.js"
             ],
             "run_at": "document_end"
         }
     ],
     "background": {
         "scripts": [
-            "src/assets/js/web-storage-object.js",
-            "src/assets/js/state.js"
+            "assets/js/web-storage-object.js",
+            "lib/state.js"
         ],
         "persistent": true
     },
     "web_accessible_resources": [
-        "src/assets/img/clippy.map.png"
+        "assets/img/clippy.map.png"
     ],
     "permissions": [
         "activeTab",
