+++ package.json
@@ -2,12 +2,13 @@
     "name": "clippy",
     "version": "1.9.0",
     "description": "Clippy MS Word Office asistant is now back to help inside your browser!",
-    "main": "src/index.js",
+    "main": "lib/index.js",
     "scripts": {
         "test": "jest --verbose",
         "test:coverage": "jest --collect-coverage",
         "test:upload-coverage": "jest --verbose --coverage --coverageReporters=text-lcov | coveralls",
-        "build": "web-ext build --ignore-files __tests__ coverage README.md package*.json src/assets/img/screenshots",
+        "build": "npm run build:js && web-ext build --ignore-files test coverage README.md package*.json assets/img/screenshots src",
+        "build:js": "rm-cli -r lib && babel src --out-dir lib",
         "lint": "eslint --max-warnings=0 src"
     },
     "repository": {
@@ -21,12 +22,16 @@
     },
     "homepage": "https://github.com/capJavert/clippy#readme",
     "devDependencies": {
+        "@babel/cli": "^7.5.5",
+        "@babel/core": "^7.5.5",
+        "@babel/preset-env": "^7.5.5",
         "coveralls": "^3.0.6",
         "eslint": "^6.2.1",
         "eslint-config-airbnb-base": "^14.0.0",
         "eslint-plugin-import": "^2.18.2",
         "husky": "^3.0.4",
         "jest": "^24.9.0",
+        "rm-cli": "^1.4.2",
         "web-ext": "^3.1.1"
     },
     "husky": {
@@ -37,7 +42,8 @@
     "jest": {
         "collectCoverageFrom": [
             "<rootDir>/src/index.js",
-            "<rootDir>/src/assets/js/state.js"
+            "<rootDir>/src/state.js"
         ]
-    }
+    },
+    "dependencies": {}
 }
